Add pwn100 to docker-compose
Also changed all of the dockerfiles to have an unconfined seccomp profile. This will enable the player to gain a shell on the containers to use /bin/sh payloads, which is indented. I believe seccomp in this manner doesn't matter on the kube cluster side, but for local development it is important.