This is a gross fix to pwn300... but it works?
In the container when the exploit is landed, there is a permissions error of opening a new process due to /proc/#/ not being writable. This brings the program under the control of GDB, which is fine, for the challenge. Such gross. There has to be a better way! This does work, but

¯\_(ツ)_/¯